test: change class name
